查看购物车

+++ shop-product-service/src/main/java/com/wanfeng/shop/product/model/vo/CartVO.java
@@ -19,7 +19,7 @@ public class CartVO implements Serializable {
     /**
      * 购买总件数
      */
-    private Integer totalNum;
+    private Long totalNum;
 
     /**
      * 购物车总价格
